backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
gunicorn==20.1.0
mysql-connector-python==8.0.33
PyYAML==6.0.1
python-dotenv==1.0.1
